// src/shared/dataprovider/api/model/successRateSummary.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface SuccessRateSummary { 
    ClientId?: number;
    ClientName?: string;
    TotalTransactionCICO?: number;
    TotalTransactionCI?: number;
    TotalTransactionCO?: number;
    TotalSuccessCountCICO?: number;
    TotalSuccessCountCI?: number;
    TotalSuccessCountCO?: number;
    TotalPendingCountCI?: number;
    TotalPendingCountCO?: number;
    StartDate?: Date;
    EndDate?: Date;
    AverageSuccessRateCICO?: number; 
    AverageSuccessRateCI?: number;
    AverageSuccessRateCO?: number;
}