// src/shared/dataprovider/api/model/withdrawalFeeConfigureDTO.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { StatusEnum } from './statusEnum';
import { WithdrawalTypeEnum } from './withdrawalTypeEnum';

export interface WithdrawalFeeConfigureDTO { 
    Id?: number;
    ClientId?: number;
    Method?: WithdrawalTypeEnum;
    MethodDescription?: string;
    FeeOnTopPercent?: number;
    FeeOnTopFixed?: number;
    FeeAtCostPercent?: number;
    FeeAtCostFixed?: number;
    LastUpdatedDate?: Date;
    Status?: StatusEnum;
    Settlement?: string;
}